Fetch event type IDs from host VM (#21)
* Fetch event type IDs from host VM

* Mark some methods host-only, use helper to safely cast long to int

* Only consider native events

+++ substratevm/src/com.oracle.svm.jfr/src/com/oracle/svm/jfr/SubstrateJVM.java
@@ -422,7 +422,7 @@ class SubstrateJVM {
 
     @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
     public boolean isEnabled(JfrEvents event) {
-        return eventSettings[event.getId()].isEnabled();
+        return eventSettings[(int) event.getId()].isEnabled();
     }
 
     /** See {@link JVM#setThreshold}. */
